Expose pairing helpers for interaction extensions
